2023/04/20 서정현

+++ Lemo/src/main/java/kr/co/Lemo/controller/UserController.java
@@ -118,10 +118,12 @@ public class UserController {
         if(authentication != null){
             new SecurityContextLogoutHandler().logout(req,resp,authentication);
 
-            // DB remember-me 쿠키 삭제
-            JdbcTokenRepositoryImpl repository = new JdbcTokenRepositoryImpl();
-            repository.setDataSource(dataSource);
-            repository.removeUserTokens(userVO.getUsername());
+            if(userVO != null){
+                // DB remember-me 쿠키 삭제
+                JdbcTokenRepositoryImpl repository = new JdbcTokenRepositoryImpl();
+                repository.setDataSource(dataSource);
+                repository.removeUserTokens(userVO.getUsername());
+            }
         }
         String fromUri = req.getHeader("Referer");
